pci: Be more careful when checking link status
Currently we assume that once the link is up it will stay up.
Unfortunately, we live in hell world where devices can bring
the link up, take it down, shake it around a bit, throw it sideways
and eventually bring it up again.

To handle these devices change how we check for the link being up
so that we only consider it up when it's been stable for two consecutive
polls.

Signed-off-by: Oliver O'Halloran <oohall@gmail.com>